+++ app/src/ai/orchestration/remote_child.rs
@@ -51,6 +51,7 @@ impl RemoteChildLaunchConfig {
 }
 
 /// Frontend-neutral output used to launch one remote child.
+#[cfg_attr(not(feature = "tui"), allow(dead_code))]
 #[derive(Clone, Debug)]
 pub struct PreparedRemoteChildLaunch {
     pub display_name: String,
@@ -91,6 +92,7 @@ pub enum CloudAgentStartupBlocker {
     GitHubAuthRequired { message: String, auth_url: String },
 }
 
+#[cfg_attr(not(feature = "tui"), allow(dead_code))]
 impl CloudAgentStartupBlocker {
     pub fn message(&self) -> &str {
         match self {
@@ -118,6 +120,7 @@ pub enum CloudAgentStartupFailure {
     Other { message: String },
 }
 
+#[cfg_attr(not(feature = "tui"), allow(dead_code))]
 impl CloudAgentStartupFailure {
     pub fn message(&self) -> &str {
         match self {
@@ -289,6 +292,7 @@ pub(crate) fn should_disable_snapshot(ctx: &AppContext) -> bool {
 }
 
 /// Builds the Oz web URL for a server-assigned agent run ID.
+#[cfg_attr(not(feature = "tui"), allow(dead_code))]
 pub fn oz_run_url(run_id: &str) -> String {
     format!("{}/runs/{run_id}", ChannelState::oz_root_url())
 }
