+++ app/src/terminal/local_tty/mod.rs
@@ -31,7 +31,7 @@ use shell::ShellStarter;
 
 pub use self::terminal_manager::{TerminalManager, get_shell_starter};
 #[cfg(feature = "tui")]
-pub use self::terminal_manager::{TerminalManagerInit, TerminalSurfaceInit, TerminalSurfaceResult};
+pub use self::terminal_manager::{TerminalManagerInit, TerminalSurfaceResult};
 #[cfg(windows)]
 pub use self::terminal_view_adaptor::shutdown_all_pty_event_loops;
 #[cfg(all(feature = "local_tty", not(feature = "remote_tty")))]
